#5 Clinics endpoints
This implements and modifies some endpoints also add a new blutprint for a new endpoint, also modifies the mqtt handler there is still somework to make it perfect and failure tolerent but for now it delivers necessary functionalities, location attribute was omitted from the dentist model. all fully implemented endpoints of clinics are tested and they work fine without problem.